Changed pathname & implemented validation
